change Sonde class to frozen dataclass
-------------
Once a `Sonde` instance is initialized, there should be no change made to its contents. This is particularly because the `Sonde` and `SondeData` class are written keeping Level-0 data in mind, which should not be modified. Therefore, the dataclass wrapper is provided with the True value for its frozen argument.

As the `Sonde` class is now kept frozen, the straightforward assignment to `sort_index` attribute cannot be made after initialization. Therefore, now the `object.__setattr__` method is used instead.