add framework for integrating config
The idea with this commit is to allow the user to interact with halodrops only via a configuration file.

By default this file is located in the directory where the user runs `halodrops` from and is called `halodrops.cfg`. Here the file only has the data directory field, enough to test out if the framework works.

This commit sets the framework for the user to provide a path to the config file as an argument in the command line, with the argument option `--configpath` or `-c`. This is done by using the function `main`. An example of using is

```
halodrops -c="../myconfigdirectory/ac3_configuration.cfg"
```

To get the `halodrops` command working, this main function is provided as an entry point in the TOML file.

**Afterthought...**

Now, calls to other scripts should work by just adding them to the main function. Ideally, other scripts will not interact with the config file. It will be the job of the `main` function to provide inputs to the other scripts by reading the config. This might mean nesting other functions in `main`, or more functions in `__init__.py`. More arguments could also be provided for the user to call only certain scripts. Or this could be set up via the config file too. The latter is the better option because when sharing data, only the version of `halodrops` and the config file is needed. The arguments to calling `halodrops` are not really stored anywhere automatically.