add weighted_fullness fn to Sonde class
Adding this function directly to the Sonde class means that:

1. we can remove the separate qc module as it is no longer needed (all qc checks will be run in the sonde class itself)

2. the fullness without the weights is not going to be needed, so that doesn't need to be a qc function in the Sonde class

3. the weighted_fullness function is now changed so that the variable argument has a default value. This default value is now
a dictionary where the keys are variable names expected in the dataset and the values are the corresponding sampling frequencies.
Therefore, the sampling_frequency argument is no longer needed.

4. This also means that the separate RD41 config file can be removed.
Reading through it will no longer be used. Instead if the user wants to provide different variables/sampling frequencies, they can
do so using the main config file, which through the main() function in halodrops.__init__ will be replaced as a non-default value
for this function.

5. Similarly, for the timestamp_frequency, the default value is now given as 4. The user can change it if they want.
The value of 4 is no longer fixed in the function.

6. Docstrings have been changed to reflect the changes in the function