check launch detect before running QC
The QC functions will fail if there is no aspen_ds detected. An
aspen_ds will not be detected if a sonde has not detected a launch.

Currently, the iterate_Sonde_method in the pipeline runs functions for all
sondes present in the provided dictionary. This concept works well but some
type of a filter is needed before the QC functions are run.

Therefore, the Sonde class now has a function which checks for launch detect.
This function becomes the first in the QC step in the pipeline dictionary.
The iterate_Sonde_method function also needs to change because when the launch
detect returns a None object, this should not propagate into the new dictionary.

Thus adding the filter_no_launch_detect function to the QC step in the pipeline
ensures that for any subsequent steps, the sondes with no launch detects are removed.