add default value to filter_flags (#64)
This commit introduces significant enhancements to the `qc_filter` function in the `sonde.py` file. The function now accepts `filter_flags` as an optional parameter, which can be a string, a list, or None. This change provides greater flexibility in specifying which QC-related attributes should be checked.

If `filter_flags` is None, all QC attributes will be checked. If it's a string, it's treated as a comma-separated list of individual flag values or prefixes of flag values. If it's a list, it's treated as a list of individual flag values or prefixes of flag values.

The function now also supports the 'all_except_<prefix>' syntax in `filter_flags`. This allows all flags to be filtered except those starting with '<prefix>'. If 'all_except_<prefix>' is provided, it should be the only value in `filter_flags`.

The function raises a ValueError if a flag in `filter_flags` does not exist as an attribute of the sonde object, or if 'all_except_<prefix>' is provided in `filter_flags` along with other values.

This commit thus makes it possible for the user to customize which QC flags will be used for filtering out before L2 and the remaining that will be propagated into L2 with their flag values.