add convert_to_si fn
this function will get the helper functions and convert to SI the variables provided as an argument.

the concept of an attribute `_interim_l2_ds` also starts with this commit. The purpose of this attribute is to carry forward changes from L1 to L2 as an interim product, such that none of the previous products are affected. This same product will keep getting updated as it goes through the different steps necessary to go from L1 to L2. The prefixed underscore is an indication that this attribute is for internal use only and not for users to access. The attribute is subject to change depending on the different steps taken. Hence, any step from L1 to L2 will check first if the attribute already exists. If it does, the attribute is overwritten after the change. If the attribute does not already exist, the starting point for this attribute is the `aspen_ds` attribute. So, the `_interim_l2_ds` attribute will be set to its starting value, i.e. `self.aspen_ds`.