add get_flight_attributes method to Sonde class
This commit introduces a new method, get_flight_attributes, to the Sonde class. This method reads flight attributes from the A-file (`self.afile`) and adds them as attributes to the sonde object. The flight attributes to be read and their corresponding renamed attribute names for the L2 file are defined in a dictionary, l2_flight_attributes_map.

The l2_flight_attributes_map dictionary has been added to the helper module. It maps the original attribute names in the A-file to the new attribute names to be used in the L2 file.